Updated DrawCard struct
